% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allmethods.R
\docType{data}
\name{allmethods}
\alias{allmethods}
\title{allmethods}
\format{
An object of class \code{data.frame} with 164 rows and 2 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/26321463}
}
\usage{
allmethods
}
\description{
The dataset include 164 normalization methods/strategies.
}
\keyword{datasets}
