% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loplot_v1.R
\name{loplot}
\alias{loplot}
\title{loplot provide the visible figure of QC-RLS correction.}
\usage{
loplot(x, z, i)
}
\arguments{
\item{x}{the file before QC-RLS correction.}

\item{z}{the file after QC-RLS correction.}

\item{i}{a index for the name of variable.}
}
\description{
loplot provide the visible figure of QC-RLS correction.
}
