% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiclass-IS.R
\name{normulticlassisall}
\alias{normulticlassisall}
\title{Multi-class (N>1) Metabolomic Study with dataset with Internal Standards (ISs).}
\usage{
normulticlassisall(fileName, IS)
}
\arguments{
\item{fileName}{Allows the user to indicate the NAME of peak table resulted from PrepareInuputFiles() (default = null).}

\item{IS}{Allows the user to indicate the column number(s) where the internal standard(s) locate (default = null)
If there is only one internal standard (IS), the column number of this IS should be listed
If there are multiple ISs, the column numbers of all ISs should be listed and separated using comma
For example, the value of argument IS that is set to “2,6,8,n” indicates that the metabolites in the 3rd, 7th, 9th,
and (n+1)th columns of your input peak table should be considered to be the IS metabolites.}
}
\description{
this function enables the performance assessment of metabolomic data processing
for multi-class dataset (with internal standards but without quality control sample) using
four criteria, and can scan thousands of processing workflows and rank them based on their performances.
}
\examples{
library(NOREVA)
\donttest{multi_is_data <- PrepareInuputFiles(dataformat = 1,
rawdata = "Multiclass_with_IS.csv")}
\donttest{normulticlassisall(fileName = multi_is_data, IS = "3,4,5")}
}
