% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectmulti_is.R
\name{normulticlassispart}
\alias{normulticlassispart}
\title{Multi-class (N>1) Metabolomic Study with dataset with Internal Standards (ISs).}
\usage{
normulticlassispart(fileName, IS, selectedMethods)
}
\arguments{
\item{fileName}{Allows the user to indicate the NAME of peak table resulted from PrepareInuputFiles() (default = null).}

\item{IS}{Allows the user to indicate the column number(s) where the internal standard(s) locate (default = null).
If there is only one internal standard (IS), the column number of this IS should be listed.
If there are multiple ISs, the column numbers of all ISs should be listed and separated using comma.
For example, the value of argument IS that is set to “2,6,8,n” indicates that the metabolites in the 3rd, 7th, 9th, and (n+1)th columns of your input peak table should be considered to be the IS metabolites.}

\item{selectedMethods}{Allows the user to indicate the NAME of the file containing the
customized workflows selected by user. The file should be in a .csv format,
and the exemplar files are provided in the NOREVA R package and available for
download at https://idrblab.org/noreva/NOREVA_exampledata.zip).}
}
\description{
this function enables the performance assessment of metabolomic
data processing for multi-class dataset (with internal standards but without quality control sample)
using four criteria, and can scan the customized workflows selected by user and rank them based on performances.
}
\examples{
library(NOREVA)
\donttest{multi_is_data <- PrepareInuputFiles(dataformat = 1,
rawdata = "Multiclass_with_IS.csv")}
\donttest{normulticlassispart(fileName = multi_is_data,
IS = "3,4,5", selectedMethods = "selectedmethods.csv")}
}
