% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_multiclassdata.R
\name{normulticlassmatrix}
\alias{normulticlassmatrix}
\title{Process the dataset of multi-class (N>1) metabolomic study.}
\usage{
normulticlassmatrix(
  datatype,
  fileName,
  IS,
  impt = NULL,
  qcsn = NULL,
  trsf = NULL,
  nmal = NULL,
  nmal2 = NULL,
  nmals = NULL
)
}
\arguments{
\item{datatype}{Allows the users to specify the data type of their input peak table (default = null).
“1” denotes the multi-class metabolomic dataset without QCSs and without ISs.
“2” denotes the multi-class metabolomic dataset with QCSs but without ISs.
“3” denotes the multi-class metabolomic dataset with ISs but without QCSs.}

\item{fileName}{Allows the users to indicate the NAME of peak table resulted from PrepareInuputFiles() (default = null)}

\item{IS}{Allows the user to indicate the column number(s) where the internal standard(s) locate (default = null).
If there is only one internal standard (IS), the column number of this IS should be listed.
If there are multiple ISs, the column numbers of all ISs should be listed and separated using comma.
For example, the value of argument IS that is set to “2,6,8,n” indicates
that the metabolites in the 3rd, 7th, 9th, and (n+1)th columns of your
input peak table should be considered to be the IS metabolites.}

\item{impt}{Input the name of imputation methods.
If set 1, method of column mean imputation.
If set 2, method of column median imputation.
If set 3, method of half of the minimum positive value.
If set 4, method of KNN imputation.}

\item{qcsn}{Input the name of qc sample correction methods.
If set 1, method of NWE (Nadaraya-Watson estimator).
If set 2, method of LLR (local linear regression).
If set 3, method of LPF (local polynomial fits).}

\item{trsf}{Input the name of transformation methods.
If set 1, method of cube root transformation.
If set 2, method of log transformation.
If set 3, none transformation method.}

\item{nmal}{Allows the users to specify the NAME of the normalization method (default = null)}

\item{nmal2}{Allows the users to specify the NAME of the normalization method (default = null)}

\item{nmals}{Allows the users to specify the NAME of the IS-based normalization method (default = null)}
}
\description{
based on a particular processing workflow (especially the one identified as well-performing
for the studied metabolomic multi-class dataset),
this function outputs the resulting levels of all metabolites among all samples
after the data processing based on that workflow. Quality control sample (QCS) and internal standard (IS)
are considered in this function.
}
\examples{
nordata <- normulticlassmatrix(datatype = 1,
fileName = multi_non_data, impt = 1, trsf = 1, nmal = 1, nmal2 = 1)
}
