% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pMulticlass-NONE.R
\name{normulticlassnoall}
\alias{normulticlassnoall}
\title{Multi-class (N>1) Metabolomic Study with dataset without QCSs and ISs.}
\usage{
normulticlassnoall(fileName, SAalpha="Y", SAbeta="Y", SAgamma="Y")
}
\arguments{
\item{fileName}{Allows the user to indicate the NAME of peak table resulted from PrepareInuputFiles() (default = null).}

\item{SAalpha}{Allows the user to specify whether the input peak table satisfies the study assumption Alpha (SAalpha, all metabolites are assumed to be equally important) (default = “Y”).
“Y” denotes that the peak table satisfies the study assumption Alpha (SAalpha).
“N” denotes that the peak table does not satisfy the study assumption Alpha (SAalpha).}

\item{SAbeta}{Allows the user to specify whether the input peak table satisfies the study assumption Beta (SAbeta, the level of metabolite abundance is constant among all samples) (default = “Y”).
“Y” denotes that the peak table satisfies the study assumption Beta (SAbeta).
“N” denotes that the peak table does not satisfy the study assumption Beta (SAbeta).}

\item{SAgamma}{Allows the user to specify whether the input table satisfies study assumption Gamma (SAγ, the intensities of most metabolites are not changed under the studied conditions) (default = “Y”).
“Y” denotes that the peak table satisfies the study assumption Gamma (SAγ).
“N” denotes that the peak table does not satisfy the study assumption Gamma (SAγ).}
}
\description{
This function handles (1) normalizing the multi-class metabolomic
data without QCSs and ISs using 168 methods/strategies,
(2) evaluating the normalization performances from multiple perspectives,
and (3) enabling the systematic comparison among all methods/strategies
based on a comprehensive performance ranking.
}
\examples{
library(NOREVA)
\donttest{multi_non_data <- PrepareInuputFiles(dataformat = 1,
rawdata = "Multiclass_without_QCSIS.csv")}
\donttest{normulticlassnoall(fileName = multi_non_data,
SAalpha="Y", SAbeta="Y", SAgamma="Y")}
}
