% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectmulti_non.R
\name{normulticlassnopart}
\alias{normulticlassnopart}
\title{Multi-class (N>1) Metabolomic Study with dataset without QCSs and ISs.}
\usage{
normulticlassnopart(fileName, selectedMethods)
}
\arguments{
\item{fileName}{Allows the user to indicate the NAME of peak table resulted from PrepareInuputFiles() (default = null).}

\item{selectedMethods}{Allows the user to indicate the NAME of the file containing the
customized workflows selected by user. The file should be in a .csv format,
and the exemplar files are provided in the NOREVA R package and available for
download at https://idrblab.org/noreva/NOREVA_exampledata.zip).}
}
\description{
this function enables performance assessment of metabolomic data
processing for multi-class dataset (without quality control sample and
without internal standard) using four criteria, and can scan the customized workflows selected
by user and rank them based on performances.
}
\examples{
library(NOREVA)
\donttest{multi_non_data <- PrepareInuputFiles(dataformat = 1,
rawdata = "Multiclass_without_QCSIS.csv")}
\donttest{normulticlassnopart(fileName = multi_non_data, selectedMethods = "selectedmethods.csv")}
}
