% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pGS-Multiclass-QCS.R
\name{normulticlassqcallgs}
\alias{normulticlassqcallgs}
\title{Multi-class (N>1) Metabolomic Study with dataset with Quality Control Samples (QCSs) and the corresponding data of golden standards for performance evaluation using Criterion e.}
\usage{
normulticlassqcallgs(fileName, GS, SAalpha="Y", SAbeta="Y", SAgamma="Y")
}
\arguments{
\item{fileName}{Allows the user to indicate the NAME of peak table resulted from PrepareInuputFiles() (default = null).}

\item{GS}{Allows the user to indicate the name of the file that contains the spike-in compounds (default = null).
The file should be in a .csv format, which provides the concentrations of spike-in compounds.}

\item{SAalpha}{Allows the user to specify whether the input peak table satisfies the study assumption Alpha (SAalpha, all metabolites are assumed to be equally important) (default = “Y”).
“Y” denotes that the peak table satisfies the study assumption Alpha (SAalpha).
“N” denotes that the peak table does not satisfy the study assumption Alpha (SAalpha).}

\item{SAbeta}{Allows the user to specify whether the input peak table satisfies the study assumption Beta (SAbeta, the level of metabolite abundance is constant among all samples) (default = “Y”).
“Y” denotes that the peak table satisfies the study assumption Beta (SAbeta).
“N” denotes that the peak table does not satisfy the study assumption Beta (SAbeta).}

\item{SAgamma}{Allows the user to specify whether the input table satisfies study assumption Gamma (SAγ, the intensities of most metabolites are not changed under the studied conditions) (default = “Y”).
“Y” denotes that the peak table satisfies the study assumption Gamma (SAγ).
“N” denotes that the peak table does not satisfy the study assumption Gamma (SAγ).}
}
\description{
this function enables the performance assessment of metabolomic data processing
for multi-class dataset (with quality control sample but without internal standard)
using five independent criteria, and can comprehensively scan thousands of processing
workflows and rank all these workflows based on their performances (assessed from four different perspectives).
}
\examples{
library(NOREVA)
\donttest{multi_qc_data <- PrepareInuputFiles(dataformat = 1,
rawdata = "Multiclass_with_QCS.csv")}
\donttest{normulticlassqcallgs(fileName = multi_qc_data,
GS = "Multiclass_with_QCS_GoldenStandard.csv", SAalpha="Y", SAbeta="Y", SAgamma="Y")}
}
