% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectmulti_qcs.R
\name{normulticlassqcpart}
\alias{normulticlassqcpart}
\title{Multi-class (N>1) Metabolomic Study with dataset with with Quality Control Samples (QCSs).}
\usage{
normulticlassqcpart(fileName, selectedMethods)
}
\arguments{
\item{fileName}{Allows the user to indicate the NAME of peak table resulted from PrepareInuputFiles() (default = null).}

\item{selectedMethods}{Allows the user to indicate the NAME of the file containing the
customized workflows selected by user. The file should be in a .csv format,
and the exemplar files are provided in the NOREVA R package and available for
download at https://idrblab.org/noreva/NOREVA_exampledata.zip).}
}
\description{
this function enables performance assessment of metabolomic data
processing for multi-class dataset (with quality control sample but without internal standard)
using four criteria, and can scan the customized workflows selected by user and rank them based on their performances.
}
\examples{
library(NOREVA)
\donttest{multi_qc_data <- PrepareInuputFiles(dataformat = 1,
rawdata = "Multiclass_with_QCS.csv")}
\donttest{normulticlassqcpart(fileName = multi_qc_data, selectedMethods = "selectedmethods.csv")}
}
