% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS-Timecourse-IS.R
\name{nortimecourseisallgs}
\alias{nortimecourseisallgs}
\title{Time-course Metabolomic Study with dataset with Internal Standards (ISs) and the corresponding data of golden standards for performance evaluation using Criterion e.}
\usage{
nortimecourseisallgs(fileName, IS, GS)
}
\arguments{
\item{fileName}{Allows the user to indicate the NAME of peak table resulted from PrepareInuputFiles() (default = null).}

\item{IS}{Allows the user to indicate the column number(s) where the internal standard(s) locate (default = null)
If there is only one internal standard (IS), the column number of this IS should be listed
If there are multiple ISs, the column numbers of all ISs should be listed and separated using comma
For example, the value of argument IS that is set to “2,6,8,n” indicates that the metabolites in the 3rd, 7th, 9th, and (n+1)th columns of your input peak table should be considered to be the IS metabolites.}

\item{GS}{Allows the user to indicate the name of the file that contains the spike-in compounds (default = null).
The file should be in a .csv format, which provides the concentrations of spike-in compounds.}
}
\description{
this function enables the performance assessment of
metabolomic data processing for time-course dataset (with internal standards but without quality control sample)
using five criteria, and can scan thousands of processing workflows and rank them based on their performances.
}
\examples{
library(NOREVA)
\donttest{timec_is_data <- PrepareInuputFiles(dataformat = 1,
rawdata = "Timecourse_with_IS.csv")}
\donttest{nortimecourseisallgs(fileName = timec_is_data,
GS = "Timecourse_with_IS_GoldenStandard.csv", IS = "4,5")}
}
