% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_timecoursedata.R
\name{nortimecoursematrix}
\alias{nortimecoursematrix}
\title{Process the dataset of time-course metabolomic study.}
\usage{
nortimecoursematrix(
  datatype,
  fileName,
  IS,
  impt = NULL,
  qcsn = NULL,
  trsf = NULL,
  nmal = NULL,
  nmal2 = NULL,
  nmals = NULL
)
}
\arguments{
\item{datatype}{Input the number of data type.
If set 1, the dataset of time-course metabolomic study without QCSs and ISs.
If set 2, the dataset of time-course metabolomic study with quality control samples (QCSs).
If set 3, the dataset of time-course metabolomic study with dataset with internal standards (ISs).}

\item{fileName}{Please find the detail information of the file format from those six sample
files in the working directory (in github) “idrblab/NOREVA/data”}

\item{IS}{Input the Column of Internal Standards. For example,
the replacement of IS to 2,6,9,n indicates that the metabolites in the 2st, 6th, 9th, and nth columns of in
your input dataset Input-Dataset.csv should be considered as the ISs or quality control metabolites.
If there is only one IS, the column number of this IS should be listed
If there are multiple ISs, the column number of all ISs should be listed and
separated by comma (,)}

\item{impt}{Input the name of imputation methods.
If set 1, method of column mean imputation.
If set 2, method of column median imputation.
If set 3, method of half of the minimum positive value.
If set 4, method of KNN imputation.}

\item{qcsn}{Input the name of qc sample correction methods.
If set 1, method of NWE (Nadaraya-Watson estimator).
If set 2, method of LLR (local linear regression).
If set 3, method of LPF (local polynomial fits).}

\item{trsf}{Input the name of transformation methods.
If set 1, method of cube root transformation.
If set 2, method of log transformation.
If set 3, none transformation method.}

\item{nmal}{Allows the users to specify the NAME of the normalization method (default = null)}

\item{nmal2}{Allows the users to specify the NAME of the normalization method (default = null)}

\item{nmals}{Allows the users to specify the NAME of the IS-based normalization method (default = null)}
}
\description{
based on a particular processing workflow
(especially the one identified as well-performing for the studied metabolomic time-course dataset),
this function outputs the resulting levels of all metabolites among all samples after the
data processing based on that workflow.
}
\examples{
nordata <- nortimecoursematrix(datatype = 1,
fileName = timec_non_data, impt = 1, trsf = 1, nmal = 1, nmal2 = 1)
}
