% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pTimecourse-NONE.R
\name{nortimecoursenoall}
\alias{nortimecoursenoall}
\title{Time-course Metabolomic Study with dataset without QCSs and ISs.}
\usage{
nortimecoursenoall(fileName, SAalpha="Y", SAbeta="Y", SAgamma="Y")
}
\arguments{
\item{fileName}{Allows the user to indicate the NAME of peak table resulted from PrepareInuputFiles() (default = null).}

\item{SAalpha}{Allows the user to specify whether the input peak table satisfies the study assumption Alpha (SAalpha, all metabolites are assumed to be equally important) (default = “Y”).
“Y” denotes that the peak table satisfies the study assumption Alpha (SAalpha).
“N” denotes that the peak table does not satisfy the study assumption Alpha (SAalpha).}

\item{SAbeta}{Allows the user to specify whether the input peak table satisfies the study assumption Beta (SAbeta, the level of metabolite abundance is constant among all samples) (default = “Y”).
“Y” denotes that the peak table satisfies the study assumption Beta (SAbeta).
“N” denotes that the peak table does not satisfy the study assumption Beta (SAbeta).}

\item{SAgamma}{Allows the user to specify whether the input table satisfies study assumption Gamma (SAγ, the intensities of most metabolites are not changed under the studied conditions) (default = “Y”).
“Y” denotes that the peak table satisfies the study assumption Gamma (SAγ).
“N” denotes that the peak table does not satisfy the study assumption Gamma (SAγ).}
}
\description{
this function enables the performance assessment of metabolomic data processing for time-course dataset
(without quality control sample and without internal standard) using four
independent criteria, and can comprehensively scan thousands of processing workflows and rank
all these workflows based on their performances (assessed from four different perspectives).
}
\examples{
library(NOREVA)
\donttest{timec_non_data <- PrepareInuputFiles(dataformat = 1,
rawdata = "Timecourse_without_QCSIS.csv")}
\donttest{nortimecoursenoall(fileName = timec_non_data,
SAalpha="Y", SAbeta="Y", SAgamma="Y")}
}
