% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pGS-Timecourse-NONE.R
\name{nortimecoursenoallgs}
\alias{nortimecoursenoallgs}
\title{Time-course Metabolomic Study with dataset without QCSs & ISs and the corresponding data of golden standards for performance evaluation using Criterion e.}
\usage{
nortimecoursenoallgs(fileName, GS, SAalpha="Y", SAbeta="Y", SAgamma="Y")
}
\arguments{
\item{fileName}{Allows the user to indicate the NAME of peak table resulted from PrepareInuputFiles() (default = null).}

\item{GS}{Allows the user to indicate the name of the file that contains the spike-in compounds (default = null).
The file should be in a .csv format, which provides the concentrations of spike-in compounds.}

\item{SAalpha}{Allows the user to specify whether the input peak table satisfies the study assumption Alpha (SAalpha, all metabolites are assumed to be equally important) (default = “Y”).
“Y” denotes that the peak table satisfies the study assumption Alpha (SAalpha).
“N” denotes that the peak table does not satisfy the study assumption Alpha (SAalpha).}

\item{SAbeta}{Allows the user to specify whether the input peak table satisfies the study assumption Beta (SAbeta, the level of metabolite abundance is constant among all samples) (default = “Y”).
“Y” denotes that the peak table satisfies the study assumption Beta (SAbeta).
“N” denotes that the peak table does not satisfy the study assumption Beta (SAbeta).}

\item{SAgamma}{Allows the user to specify whether the input table satisfies study assumption Gamma (SAγ, the intensities of most metabolites are not changed under the studied conditions) (default = “Y”).
“Y” denotes that the peak table satisfies the study assumption Gamma (SAγ).
“N” denotes that the peak table does not satisfy the study assumption Gamma (SAγ).}
}
\description{
this function enables the performance assessment of metabolomic data processing
for time-course dataset (without quality control sample and without internal standard)
using five criteria, and can scan thousands of workflows and rank them based on their performances.
}
\examples{
library(NOREVA)
\donttest{timec_non_data <- PrepareInuputFiles(dataformat = 1,
rawdata = "Timecourse_without_QCSIS.csv")}
\donttest{nortimecoursenoallgs(fileName = timec_non_data,
GS = "Timecourse_without_QCSIS_GoldenStandard.csv", SAalpha="Y", SAbeta="Y", SAgamma="Y")}
}
