% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecttimec_qcs.R
\name{nortimecourseqcpart}
\alias{nortimecourseqcpart}
\title{Time-course Metabolomic Study with dataset with QCSs.}
\usage{
nortimecourseqcpart(fileName, selectedMethods)
}
\arguments{
\item{fileName}{Allows the user to indicate the NAME of peak table resulted from PrepareInuputFiles() (default = null).}

\item{selectedMethods}{Allows the user to indicate the NAME of the file containing the
customized workflows selected by user. The file should be in a .csv format,
and the exemplar files are provided in the NOREVA R package and available for
download at https://idrblab.org/noreva/NOREVA_exampledata.zip).}
}
\description{
this function enables performance assessment of metabolomic
data processing for time-course data (with quality control sample but without internal standard)
using four criteria, and can scan the customized workflows selected by user
and rank them based on their performances.
}
\examples{
library(NOREVA)
\donttest{timec_qc_data <- PrepareInuputFiles(dataformat = 1,
rawdata = "Timecourse_with_QCS.csv")}
\donttest{nortimecourseqcpart(fileName = timec_qc_data,
selectedMethods = "selectedmethods.csv")}
}
