% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runjava.R
\name{norvisualization}
\alias{norvisualization}
\title{Plotcircularplot Metabolomic Study with dataset without QCSs and ISs.}
\usage{
norvisualization(data, outputfile,cutoff="100",outputtype,maxValue,
colorSet, totalAngle,bgColor, fontColor)
}
\arguments{
\item{data}{String, the input file name}

\item{outputfile}{Format string, together with cutoff value and type to generate formatted file name}

\item{cutoff}{Integer, which means to filter the results, the default is 100}

\item{outputtype}{Double-precision floating-point number, representing the characteristic value represented by the maximum length of the rectangle, the default is 40}

\item{maxValue}{Double-precision floating-point number, representing the characteristic value represented by the maximum length of the rectangle, the default is 40}

\item{colorSet}{Hexadecimal color string group, representing the four-layer color setting of the graph from the inside to the outside, the default is red (#EA4335), blue (#4285F4), yellow (#FBBC05), purple (#800080)}

\item{totalAngle}{Double-precision floating-point number, representing the total angle of rotation of the drawing, in degrees, the default value is 340}

\item{bgColor}{Hexadecimal color string, representing the background color of the graphic drawing, the default is white (#FFFFFF).}

\item{fontColor}{Hexadecimal color string, representing the font color, the default is black (#000000).}
}
\description{
plot circular plot
}
\examples{
library(NOREVA)
\donttest{multi_qc_data <- PrepareInuputFiles(dataformat = 1,
rawdata = "Multiclass_with_QCS.csv")}
\donttest{normulticlassqcall(fileName = multi_qc_data,
assum_a="Y", assum_b="Y", assum_c="N")}
\donttest{norvisualization(data = "OUTPUT-NOREVA-Overall.Ranking.Data.csv",
cutoff = "100")}
}
