% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareinput.R
\name{PrepareInuputFiles}
\alias{PrepareInuputFiles}
\title{The preparation and input of peak table}
\usage{
PrepareInuputFiles(dataformat, rawdata, label)
}
\arguments{
\item{dataformat}{Allows the user to specify the FORMAT of their input peak table (default = null)
“1” denotes a standardized format of peak table accepted by NOREVA.
“2” denotes the customized formats of peak table generated by 12 popular tools (such as XCMS).}

\item{rawdata}{Allows the user to indicate the NAME of their input peak table file (default = null).}

\item{label}{Allows the user to indicate the NAME of their input label file (default = null).}
}
\value{
prepareinputfiles matrix
}
\description{
this function enables the preparation and input of peak table
which facilitate the subsequent application of other NOREVA functions.
}
\examples{
library(NOREVA)
\donttest{multi_qcs_data <- PrepareInuputFiles(dataformat = 1,
rawdata = "Multiclass_with_QCS.csv")}
\donttest{my_test <- PrepareInuputFiles(dataformat = 2,
rawdata = "example_OpenMS.csv", label = "OpenMS_Label.txt")}
}
