% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loplot_v1.R
\name{loplot}
\alias{loplot}
\title{Loplot}
\usage{
loplot(x,z,i)
}
\arguments{
\item{x}{the file before QC-RLS correction.}

\item{z}{the file after QC-RLS correction.}

\item{i}{a index for the name of variable.}
}
\description{
Loplot, a visible figure of signal correction loplot
provide the visible figure of QC-based correction. See the details at the following References.
}
\references{
statTarget: a streamlined tool for signal drift correction
and interpretations of quantitative mass spectrometry-based omics data.
Luan H, Ji F, Chen Y, Cai Z. 2018, Analytica Chimica Acta.
}
